requirements.txt
opencv-python
Pillow
flask